Make it possible to flip the avatar
The plan is to have my real photo on the other side of the avatar. I
don't have a nice photo of myself, so we just duplicate the image so far.

This commit also includes a bunch of SASS files reorganizations. I
decided to put `fonts.css.sass` under main for consistency, too. Oh, and
we use mixins from now on.